requirements.txt
keyboard==0.13.5
pywin32==306
pystray==0.19.5
Pillow==10.2.0
pyinstaller==6.3.0
